Use 64 bit settings for default table sizes, fix swapped config setting.

+++ include/bitcoin/database/settings.hpp
@@ -26,6 +26,7 @@
 namespace libbitcoin {
 namespace database {
 
+/// 32 bit builds will warn on size downcasts.
 /// TODO: helper methods to embed/construct full paths (see store()).
 /// Common database configuration settings, properties not thread safe.
 struct BCD_API settings
@@ -42,46 +43,46 @@ struct BCD_API settings
     /// -----------------------------------------------------------------------
 
     uint32_t header_buckets;
-    uint32_t header_size;
+    uint64_t header_size;
     uint16_t header_rate;
 
     uint32_t point_buckets;
-    uint32_t point_size;
+    uint64_t point_size;
     uint16_t point_rate;
 
     uint32_t input_buckets;
-    uint32_t input_size;
+    uint64_t input_size;
     uint16_t input_rate;
 
-    uint32_t output_size;
+    uint64_t output_size;
     uint16_t output_rate;
 
-    uint32_t puts_size;
+    uint64_t puts_size;
     uint16_t puts_rate;
 
     uint32_t tx_buckets;
-    uint32_t tx_size;
+    uint64_t tx_size;
     uint16_t tx_rate;
 
     uint32_t txs_buckets;
-    uint32_t txs_size;
+    uint64_t txs_size;
     uint16_t txs_rate;
 
     /// Indexes.
     /// -----------------------------------------------------------------------
 
     uint32_t address_buckets;
-    uint32_t address_size;
+    uint64_t address_size;
     uint16_t address_rate;
 
-    uint32_t candidate_size;
+    uint64_t candidate_size;
     uint16_t candidate_rate;
 
-    uint32_t confirmed_size;
+    uint64_t confirmed_size;
     uint16_t confirmed_rate;
 
     uint32_t strong_tx_buckets;
-    uint32_t strong_tx_size;
+    uint64_t strong_tx_size;
     uint16_t strong_tx_rate;
 
     /// Caches.
@@ -91,19 +92,19 @@ struct BCD_API settings
     uint16_t bootstrap_rate;
 
     uint32_t buffer_buckets;
-    uint32_t buffer_size;
+    uint64_t buffer_size;
     uint16_t buffer_rate;
 
     uint32_t neutrino_buckets;
-    uint32_t neutrino_size;
+    uint64_t neutrino_size;
     uint16_t neutrino_rate;
 
     uint32_t validated_bk_buckets;
-    uint32_t validated_bk_size;
+    uint64_t validated_bk_size;
     uint16_t validated_bk_rate;
 
     uint32_t validated_tx_buckets;
-    uint32_t validated_tx_size;
+    uint64_t validated_tx_size;
     uint16_t validated_tx_rate;
 };
